# backend/cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_http_stock_movement.CreateStockMovementRequest:
    properties:
      action:
        type: string
      date:
        type: string
      quantity:
        type: integer
      reason:
        type: string
      source:
        type: string
    required:
    - action
    - date
    - quantity
    type: object
host: localhost:8080
info:
  contact: {}
  description: Stock management application API.
  title: Stockify API
  version: "1.0"
paths:
...
  /product/{id}/stock-movements/:
    get:
      description: Retrieve all stock movements for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: list of movements
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get stock movements by product
      tags:
      - Stock Movement
    post:
      consumes:
      - application/json
      description: Record a stock movement (RESTOCK, REFUND, SOLD, BROKEN) for a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Stock movement data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_stock_movement.CreateStockMovementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Create a stock movement
      tags:
      - Stock Movement
  /product/{id}/stock-movements/summary:
    get:
      description: Get in/out summary for a product, optionally filtered by date
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Date filter: 1d, 1w, 1m, 3m, 6m, 1y'
        in: query
        name: dateFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: summary data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id or filter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get stock movement summary by product
      tags:
      - Stock Movement
...
  /stock-movements/:
    get:
      description: Get overall in/out summary across all products, optionally filtered
        by date
      parameters:
      - description: 'Date filter: 1d, 1w, 1m, 3m, 6m, 1y'
        in: query
        name: dateFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: global summary
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid filter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get global stock movement summary
      tags:
      - Stock Movement
  /stock-movements/top-movers:
    get:
      description: Get products with highest stock out activity, optionally filtered
        by date
      parameters:
      - description: Max results (default 10)
        in: query
        name: limit
        type: integer
      - description: 'Date filter: 1d, 1w, 1m, 3m, 6m, 1y'
        in: query
        name: dateFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: list of top movers
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid filter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get top movers
      tags:
      - Stock Movement
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"